feat: add randomAttack function
