+++ src/service/Game/attack.ts
@@ -3,7 +3,13 @@ import { games } from "../../db/game";
 import { Attack, Field, Player } from "../../utils/type/interface";
 import { turnPlayer } from "./turnPlayer";
 import { showPlayersAttack } from "./showPlayersAttack";
-import { announceWinner } from "./announceWinner";
+// import { announceWinner } from "./announceWinner";
+
+import { showWinners } from "./showWinners";
+import { rooms } from "../../db/rooms";
+import { createJsonMessage } from "../../utils/utils";
+import { users } from "../../db/player";
+import { removeGame } from "../../db/game";
 
 export const attack = (user: Player, dataAttack: Attack) => {
   const { x, y, gameId, indexPlayer } = dataAttack;
@@ -44,12 +50,10 @@ export const attack = (user: Player, dataAttack: Attack) => {
   if (isHit) {
     cell.leftSide--;
     if (cell.leftSide === 0) {
-      // Уничтожение корабля
-      --opponentPlayer.shipsLeft; // Правильное уменьшение количества кораблей
+      --opponentPlayer.shipsLeft;
       processDestroyedShip(cell, gameId, currentPlayer.index);
 
       if (opponentPlayer.shipsLeft === 0) {
-        // Объявление победителя, если это был последний корабль
         announceWinner(gameId, currentPlayer.index);
         return;
       }
@@ -59,15 +63,17 @@ export const attack = (user: Player, dataAttack: Attack) => {
         turnPlayer(nextPlayerIndex, gameId);
       }
     } else {
-      // Попадание, но корабль ещё не уничтожен
       showPlayersAttack(gameId, "attack", {
         position: { y, x },
         currentPlayer: currentPlayer.index,
         status: "shot",
       });
+      const nextPlayerIndex = game.roomUsers.find((ru) => ru.index !== +indexPlayer)?.index;
+      if (nextPlayerIndex !== undefined) {
+        turnPlayer(nextPlayerIndex, gameId);
+      }
     }
   } else {
-    // Обработка промаха
     showPlayersAttack(gameId, "attack", {
       position: { y, x },
       currentPlayer: currentPlayer.index,
@@ -80,15 +86,7 @@ export const attack = (user: Player, dataAttack: Attack) => {
   }
 };
 
-function processDestroyedShip(cell: Field, gameId: string, currentPlayerIndex: number) {
-  cell.shipTheCells.forEach((shipCell) => {
-    showPlayersAttack(gameId, "attack", {
-      position: { x: shipCell[0], y: shipCell[1] },
-      currentPlayer: currentPlayerIndex,
-      status: "killed",
-    });
-  });
-
+const processDestroyedShip = (cell: Field, gameId: string, currentPlayerIndex: number) => {
   cell.overCells.forEach((aroundCell) => {
     const [x, y] = aroundCell;
     showPlayersAttack(gameId, "attack", {
@@ -97,4 +95,51 @@ function processDestroyedShip(cell: Field, gameId: string, currentPlayerIndex: n
       status: "missed",
     });
   });
-}
+
+  cell.shipTheCells.forEach((shipCell) => {
+    console.log({ x: shipCell[0], y: shipCell[1] });
+    showPlayersAttack(gameId, "attack", {
+      position: { x: shipCell[0], y: shipCell[1] },
+      currentPlayer: currentPlayerIndex,
+      status: "killed",
+    });
+  });
+};
+
+export const announceWinner = (gameId: string, winnerIndex: number) => {
+  const game = games.find((g) => g.roomId === gameId);
+  if (!game) {
+    console.error(`Game not found: ${gameId}`);
+    return;
+  }
+
+  game.roomUsers.forEach((roomUser) => {
+    const player = users.get(roomUser.name);
+    if (player) {
+      if (player.index === winnerIndex) {
+        player.winner = (player.winner || 0) + 1;
+      }
+
+      if (player.ws) {
+        player.ws.send(createJsonMessage("finish", { winPlayer: winnerIndex }));
+      }
+    }
+  });
+
+  users.forEach((player) => {
+    if (player.ws) {
+      player.ws.send(createJsonMessage("update_winners", showWinners()));
+    }
+  });
+  games.forEach((game) => {
+    game.roomId === gameId;
+    rooms.push(game);
+  });
+  removeGame(gameId);
+  
+  // games.forEach((game) => {
+  //   if (game.roomUsers.some((user) => user.index === winnerIndex)) {
+  //     game.roomUsers = game.roomUsers.filter((user) => user.index !== winnerIndex);
+  //   }
+  // });
+};
