+++ src/service/Game/showWinners.ts
@@ -2,9 +2,11 @@ import { users } from "../../db/player";
 import { Winners } from "../../utils/type/interface";
 
 export const showWinners = () => {
-  let winners: Winners[] = Array.from(users.values())
-    .filter((user) => user.winner && user.winner > 0)
-    .map((user) => ({ name: user.name || "Unknown", wins: user.winner }));
-
+  let winners: Winners[] = [];
+  users.forEach((user) => {
+    if (user.winner !== undefined && user.winner > 0) {
+      winners.push({ name: user.name || "Unknown", wins: user.winner });
+    }
+  });
   return winners;
 };
